Split types into separate files for readability
